modify: adjusted the import function

+++ app/Filament/Resources/TviResource/Pages/ListTvis.php
@@ -43,21 +43,27 @@ class ListTvis extends ListRecords
                 ->label('Import')
                 ->icon('heroicon-o-document-arrow-up')
                 ->form([
-                    FileUpload::make('attachment')
-                        ->required(),
+                    FileUpload::make('file')
+                        ->label('Import District')
+                        ->required()
+                        ->markAsRequired(false)
+                        ->disk('local')
+                        ->directory('imports')
+                        ->acceptedFileTypes(['application/vnd.openxmlformats-officedocument.spreadsheetml.sheet']),
                 ])
                 ->action(function (array $data) {
-                    $filePath = public_path('storage/' . $data['attachment']);
+                    if (isset($data['file']) && is_string($data['file'])) {
+                        $filePath = storage_path('app/' . $data['file']);
 
-                    try {
-                        Excel::import(new TviImport, $filePath);
-                        NotificationHandler::sendSuccessNotification('Import Successful', 'The institution training programs have been successfully imported from the file.');
-
-                    } catch (Exception $e) {
-                        NotificationHandler::sendErrorNotification('Import Failed', 'There was an issue importing the institution training programs: ' . $e->getMessage());
-                    } finally {
-                        if (file_exists($filePath)) {
-                            unlink($filePath);
+                        try {
+                            Excel::import(new TviImport, $filePath);
+                            NotificationHandler::sendSuccessNotification('Import Successful', 'The Institutions have been successfully imported from the file.');
+                        } catch (Exception $e) {
+                            NotificationHandler::sendErrorNotification('Import Failed', 'There was an issue importing the institutions: ' . $e->getMessage());
+                        } finally {
+                            if (file_exists($filePath)) {
+                                unlink($filePath);
+                            }
                         }
                     }
                 }),
